feat: create add club book modal component
