+++ src/components/clubBook/index.ts
@@ -1,0 +1,1 @@
+export * from './AddClubBookModal';
